% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expconf.R
\name{exp.conf}
\alias{exp.conf}
\title{Confidence Functions for an Exponential Distribution}
\usage{
\method{exp}{conf}(x, method, plot = TRUE)
}
\arguments{
\item{x}{a (non-empty) numeric vector of data values}

\item{method}{a character string specifying the method of creating the functions, must be one of "wilks" or "pivot"}

\item{plot}{a logical indicating whether you want a paired t-test}
}
\value{
list with distribution curve, confidence curve, density                curve, and quantile function
}
\description{
Computes confidence functions for exponential distributions
}
\examples{
exp.conf(x=rexp(10), method="pivot")
exp.conf(x=rexp(100), method="wilks")
}
