% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proptest.R
\name{prop.conf}
\alias{prop.conf}
\title{Confidence functions for prop Test}
\usage{
prop.conf(x, n, method = "mid", plot = TRUE, fullrange = "99.9 CI")
}
\arguments{
\item{x}{a vector of counts of successes.}

\item{n}{a vector of counts of trials.}

\item{method}{a character string specifying the method of creating the functions, must be one of "mid" (default) or "Z"}

\item{plot}{a logical indicating whether you want the plots to be generated.}

\item{fullrange}{a character string specifying the range of the plots, must be one of "99.9 CI" (default) or "full"}
}
\value{
A list containing the confidence functions.
}
\description{
Computes confidence functions for proportion tests on vectors of data.
}
\examples{
prop.conf(300, 733)
}
