% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttest.R
\name{t.confdist.summary}
\alias{t.confdist.summary}
\title{Confidence functions for t-Test with summary data}
\usage{
\method{t}{confdist.summary}(mean, sd, n, plot = TRUE)
}
\arguments{
\item{mean}{a (non-empty) numeric vector of data values.}

\item{sd}{a (non-empty) numeric vector of data values.}

\item{n}{a (non-empty) numeric vector of data values.}

\item{plot}{a logical indicating whether you want the plots to be generated.}
}
\value{
A list containing the confidence functions.
}
\description{
Computes confidence functions for one and two sample t-tests on summary data.
}
\examples{
t.confdist(c(3,5), c(1,.5), c(10, 12))
}
